requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
asyncio-mqtt==0.16.1
